Update GradeChart.js
now it works, get the user input from the webpage and create a chart with associated input numbers...
Note: since the initial array is empty, only the 1st click works, the 2nd or 3rd click will push the input to append the array. If the user likes to get a new chart by every click, we can change the initial array to [0,0,0,0,0...] and modify elements by assigning (e.g. a[i] = 10). 
I need to discuss with my teammates to determine.